Moved clamp function to SITools
